+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the Branch Web SDK will be documented here.
 The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 
 ## [VERSION] - unreleased
+### New Functionality
+- Set the app banner to either never show a second time, or wait a specified number of days before showing again
+
+### Changed
+- banner option `forgetHide` changed to `showAgain`, still backwards compatabile
 
 ## [v1.5.1] - 2015-05-13
 - Fixes IE11 bug with XHRRequests that caused an InvalidState error